Rename CpsText to Text
